requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.3.0
xgboost>=2.0.0
mlflow>=2.8.0
streamlit>=1.28.0
requests>=2.28.0
pyyaml>=6.0
plotly>=5.15.0
python-dotenv>=1.0.0
shap>=0.42.0
matplotlib>=3.7.0
seaborn>=0.12.0
